Export cmpt moved up cleaner to have editor only editor
